requirements.txt
python-telegram-bot==20.3
yt-dlp==2024.1
python-dotenv==1.0.0